backend/requirements.txt
# Core
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
gunicorn>=21.0.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.12.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib>=1.7.4
bcrypt==4.0.1
python-multipart>=0.0.6

# Configuration
pydantic-settings>=2.0.0
pydantic[email]>=2.0.0
python-dotenv>=1.0.0

# Storage
minio>=7.2.0

# Task Queue
celery>=5.3.0
redis>=5.0.0

# RAG
chromadb>=0.4.0
sentence-transformers>=2.2.0
langchain>=0.1.0
langchain-community>=0.0.10

# PDF Processing
pypdf>=3.0.0

# Security & Rate Limiting
slowapi>=0.1.9

# HTTP
requests>=2.31.0
httpx>=0.25.0

# Monitoring
prometheus-client>=0.19.0

# Utilities
python-dateutil>=2.8.0
# Production RAG Enhancements
pymupdf>=1.23.8
pdfplumber>=0.10.3
pytesseract>=0.3.10
pillow>=10.1.0
langchain-text-splitters>=0.0.1
rank-bm25>=0.2.2

# Email & Security
email-validator>=2.1.0

# Google OAuth
authlib>=1.3.0
httpx>=0.27.0
pytest>=7.4.0
